+++ core/src/main/java/de/bitbrain/braingdx/debug/DebugMetric.java
@@ -1,0 +1,5 @@
+package de.bitbrain.braingdx.debug;
+
+public interface DebugMetric {
+   String getCurrentValue();
+}