+++ core/src/main/java/de/bitbrain/braingdx/debug/ui/GameWorldPanel.java
@@ -1,0 +1,6 @@
+package de.bitbrain.braingdx.debug.ui;
+
+import com.badlogic.gdx.scenes.scene2d.ui.VerticalGroup;
+
+public class GameWorldPanel extends VerticalGroup {
+}